sends verification codes as an alternative to links to verify e-mail and phone
With this change, you should not send the verification e-mail to the contact
in the `user_verification` signal. E-mail and text message are automatically
sent through an `EmailVerificationBackend` and `PhoneVerificationBackend`
respectively.

The `user_email_verification` and `user_phone_verification` signals are only
there in case further processing is required after the verification code/link
has been sent to the channel.