fixes 500 in /api/auth/activate/{verification_key} API
When an `email_code` is present in the Contact db record,
`VerifyCompleteMixin.prefetch_contact_info` would add it to `initial_data`,
which then would propagate that `email_code` to `clean_data` and trigger
a call to `finalize_email_verification`.
As `VerifyCompleteMixin.create_user` is called after `auth_check_mfa`,
at the point we attempt to `activate_user`, there would be no record
with the verification_key URL argument in the database.