Remove call to deleted django-rest-framework compat wrapper
Django-REST-Framework authors removed this compatibility function from
the library a while ago, and we didn't need it anyway. AFAIK, it's
sole purpose was implementing the .distinct() method for old Oracle
DBMS backends, which did not support the underlying SQL command at the
time.

We don't use Oracle for this DB, so we can just Django's built-in
.distinct() for both new and legacy setups.